Struct ipp::client::IppClient

pub struct IppClient { /* fields omitted */ }

IPP client.

IPP client is responsible for sending requests to IPP server.

Methods

impl IppClient

pub fn new(uri: &str) -> IppClient

Create new instance of the client

pub fn with_root_certificates<T>(uri: &str, certfiles: &[T]) -> IppClient where
T: AsRef<str>,

Create new instance of the client with a list of root CA certificates

uri - target printer URI

certfiles - list of certificate file names

pub fn set_verify_hostname(&mut self, verify: bool)

Enable or disable host name validation for SSL transport. By default it is enabled.

pub fn send<T: IppOperation>(&self, operation: T) -> Result<IppAttributeList>

send IPP operation

pub fn send_request(
&self,
request: IppRequestResponse
) -> Result<IppRequestResponse>

Send request and return response