Struct ipp::operation::CreateJob

pub struct CreateJob {
    // some fields omitted
}

IPP operation Create-Job

Methods

impl CreateJob

fn new(uri: &str, job_name: Option<&str>) -> CreateJob

Create Create-Job operation Parameters: * uri - printer URI * job_name - optional job name (job-name)

fn set_job_attribute(&mut self, attribute: IppAttribute)

Set extra job attribute for this operation, for example colormodel=grayscale

fn execute_and_get_job_id(&mut self) -> Result<i32>

Convenience method to execute the request and return the job-id