Struct ipp::operation::PrintJob

pub struct PrintJob<'a> {
    // some fields omitted
}

IPP operation Print-Job

Methods

impl<'a> PrintJob<'a>

fn new(uri: &str, reader: &'a mut Read, user_name: &str, job_name: Option<&str>) -> PrintJob<'a>

Create Print-Job operation Parameters: * uri - printer URI * reader - std::io::Read reference which points to the data to be printed * user_name - name of the user (requesting-user-name) * job_name - optional job name (job-name)

fn set_job_attribute(&mut self, attribute: IppAttribute)

Set extra job attribute for this operation, for example colormodel=grayscale

fn execute_and_get_job_id(&mut self) -> Result<i32>