Enum ipp::value::IppValue

pub enum IppValue {

    Integer(i32),
    Enum(i32),
    OctetString(String),
    TextWithoutLanguage(String),
    NameWithoutLanguage(String),
    Charset(String),
    NaturalLanguage(String),
    Uri(String),
    UriScheme(String),
    RangeOfInteger {
        min: i32,
        max: i32,
    },
    Boolean(bool),
    Keyword(String),
    Array(Vec<IppValue>),
    Collection(Vec<IppValue>),
    MimeMediaType(String),
    DateTime {
        year: u16,
        month: u8,
        day: u8,
        hour: u8,
        minutes: u8,
        seconds: u8,
        deci_seconds: u8,
        utc_dir: char,
        utc_hours: u8,
        utc_mins: u8,
    },
    MemberAttrName(String),
    Resolution {
        cross_feed: i32,
        feed: i32,
        units: i8,
    },
    NoValue,
    Other {
        tag: u8,
        data: Bytes,
    },

}

IPP attribute values as defined in RFC 8010

Variants

Integer(i32)

Tuple Fields of Integer

0: i32

Enum(i32)

Tuple Fields of Enum

0: i32

OctetString(String)

Tuple Fields of OctetString

0: String

TextWithoutLanguage(String)

Tuple Fields of TextWithoutLanguage

0: String

NameWithoutLanguage(String)

Tuple Fields of NameWithoutLanguage

0: String

Charset(String)

Tuple Fields of Charset

0: String

NaturalLanguage(String)

Tuple Fields of NaturalLanguage

0: String

Uri(String)

Tuple Fields of Uri

0: String

UriScheme(String)

Tuple Fields of UriScheme

0: String

RangeOfInteger

Fields of RangeOfInteger

min: i32max: i32

Boolean(bool)

Tuple Fields of Boolean

0: bool

Keyword(String)

Tuple Fields of Keyword

0: String

Array(Vec<IppValue>)

Tuple Fields of Array

0: Vec<IppValue>

Collection(Vec<IppValue>)

Tuple Fields of Collection

0: Vec<IppValue>

MimeMediaType(String)

Tuple Fields of MimeMediaType

0: String

DateTime

Fields of DateTime

year: u16month: u8day: u8hour: u8minutes: u8seconds: u8deci_seconds: u8utc_dir: charutc_hours: u8utc_mins: u8

MemberAttrName(String)

Tuple Fields of MemberAttrName

0: String

Resolution

Fields of Resolution

cross_feed: i32feed: i32units: i8

NoValue

Other

Fields of Other

tag: u8data: Bytes

Implementations

impl IppValue

pub fn as_integer_mut(&mut self) -> Option<&mut i32>

Optionally returns mutable references to the inner fields if this is a IppValue::Integer, otherwise None

pub fn as_integer(&self) -> Option<&i32>

Optionally returns references to the inner fields if this is a IppValue::Integer, otherwise None

pub fn into_integer(self) -> Result<i32, Self>

Returns the inner fields if this is a IppValue::Integer, otherwise returns back the enum in the Err case of the result

pub fn as_enum_mut(&mut self) -> Option<&mut i32>

Optionally returns mutable references to the inner fields if this is a IppValue::Enum, otherwise None

pub fn as_enum(&self) -> Option<&i32>

Optionally returns references to the inner fields if this is a IppValue::Enum, otherwise None

pub fn into_enum(self) -> Result<i32, Self>

Returns the inner fields if this is a IppValue::Enum, otherwise returns back the enum in the Err case of the result

pub fn as_octet_string_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a IppValue::OctetString, otherwise None

pub fn as_octet_string(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a IppValue::OctetString, otherwise None

pub fn into_octet_string(self) -> Result<String, Self>

Returns the inner fields if this is a IppValue::OctetString, otherwise returns back the enum in the Err case of the result

pub fn as_text_without_language_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a IppValue::TextWithoutLanguage, otherwise None

pub fn as_text_without_language(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a IppValue::TextWithoutLanguage, otherwise None

pub fn into_text_without_language(self) -> Result<String, Self>

Returns the inner fields if this is a IppValue::TextWithoutLanguage, otherwise returns back the enum in the Err case of the result

pub fn as_name_without_language_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a IppValue::NameWithoutLanguage, otherwise None

pub fn as_name_without_language(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a IppValue::NameWithoutLanguage, otherwise None

pub fn into_name_without_language(self) -> Result<String, Self>

Returns the inner fields if this is a IppValue::NameWithoutLanguage, otherwise returns back the enum in the Err case of the result

pub fn as_charset_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a IppValue::Charset, otherwise None

pub fn as_charset(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a IppValue::Charset, otherwise None

pub fn into_charset(self) -> Result<String, Self>

Returns the inner fields if this is a IppValue::Charset, otherwise returns back the enum in the Err case of the result

pub fn as_natural_language_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a IppValue::NaturalLanguage, otherwise None

pub fn as_natural_language(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a IppValue::NaturalLanguage, otherwise None

pub fn into_natural_language(self) -> Result<String, Self>

Returns the inner fields if this is a IppValue::NaturalLanguage, otherwise returns back the enum in the Err case of the result

pub fn as_uri_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a IppValue::Uri, otherwise None

pub fn as_uri(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a IppValue::Uri, otherwise None

pub fn into_uri(self) -> Result<String, Self>

Returns the inner fields if this is a IppValue::Uri, otherwise returns back the enum in the Err case of the result

pub fn as_uri_scheme_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a IppValue::UriScheme, otherwise None

pub fn as_uri_scheme(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a IppValue::UriScheme, otherwise None

pub fn into_uri_scheme(self) -> Result<String, Self>

Returns the inner fields if this is a IppValue::UriScheme, otherwise returns back the enum in the Err case of the result

pub fn as_range_of_integer_mut(&mut self) -> Option<(&mut i32, &mut i32)>

Optionally returns mutable references to the inner fields if this is a IppValue::RangeOfInteger, otherwise None

pub fn as_range_of_integer(&self) -> Option<(&i32, &i32)>

Optionally returns references to the inner fields if this is a IppValue::RangeOfInteger, otherwise None

pub fn into_range_of_integer(self) -> Result<(i32, i32), Self>

Returns the inner fields if this is a IppValue::RangeOfInteger, otherwise returns back the enum in the Err case of the result

pub fn as_boolean_mut(&mut self) -> Option<&mut bool>

Optionally returns mutable references to the inner fields if this is a IppValue::Boolean, otherwise None

pub fn as_boolean(&self) -> Option<&bool>

Optionally returns references to the inner fields if this is a IppValue::Boolean, otherwise None

pub fn into_boolean(self) -> Result<bool, Self>

Returns the inner fields if this is a IppValue::Boolean, otherwise returns back the enum in the Err case of the result

pub fn as_keyword_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a IppValue::Keyword, otherwise None

pub fn as_keyword(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a IppValue::Keyword, otherwise None

pub fn into_keyword(self) -> Result<String, Self>

Returns the inner fields if this is a IppValue::Keyword, otherwise returns back the enum in the Err case of the result

pub fn as_array_mut(&mut self) -> Option<&mut Vec<IppValue>>

Optionally returns mutable references to the inner fields if this is a IppValue::Array, otherwise None

pub fn as_array(&self) -> Option<&Vec<IppValue>>

Optionally returns references to the inner fields if this is a IppValue::Array, otherwise None

pub fn into_array(self) -> Result<Vec<IppValue>, Self>

Returns the inner fields if this is a IppValue::Array, otherwise returns back the enum in the Err case of the result

pub fn as_collection_mut(&mut self) -> Option<&mut Vec<IppValue>>

Optionally returns mutable references to the inner fields if this is a IppValue::Collection, otherwise None

pub fn as_collection(&self) -> Option<&Vec<IppValue>>

Optionally returns references to the inner fields if this is a IppValue::Collection, otherwise None

pub fn into_collection(self) -> Result<Vec<IppValue>, Self>

Returns the inner fields if this is a IppValue::Collection, otherwise returns back the enum in the Err case of the result

pub fn as_mime_media_type_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a IppValue::MimeMediaType, otherwise None

pub fn as_mime_media_type(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a IppValue::MimeMediaType, otherwise None

pub fn into_mime_media_type(self) -> Result<String, Self>

Returns the inner fields if this is a IppValue::MimeMediaType, otherwise returns back the enum in the Err case of the result

pub fn as_date_time_mut(
&mut self
) -> Option<(&mut u16, &mut u8, &mut u8, &mut u8, &mut u8, &mut u8, &mut u8, &mut char, &mut u8, &mut u8)>

Optionally returns mutable references to the inner fields if this is a IppValue::DateTime, otherwise None

pub fn as_date_time(
&self
) -> Option<(&u16, &u8, &u8, &u8, &u8, &u8, &u8, &char, &u8, &u8)>

Optionally returns references to the inner fields if this is a IppValue::DateTime, otherwise None

pub fn into_date_time(
self
) -> Result<(u16, u8, u8, u8, u8, u8, u8, char, u8, u8), Self>

Returns the inner fields if this is a IppValue::DateTime, otherwise returns back the enum in the Err case of the result

pub fn as_member_attr_name_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a IppValue::MemberAttrName, otherwise None

pub fn as_member_attr_name(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a IppValue::MemberAttrName, otherwise None

pub fn into_member_attr_name(self) -> Result<String, Self>

Returns the inner fields if this is a IppValue::MemberAttrName, otherwise returns back the enum in the Err case of the result

pub fn as_resolution_mut(&mut self) -> Option<(&mut i32, &mut i32, &mut i8)>

Optionally returns mutable references to the inner fields if this is a IppValue::Resolution, otherwise None

pub fn as_resolution(&self) -> Option<(&i32, &i32, &i8)>

Optionally returns references to the inner fields if this is a IppValue::Resolution, otherwise None

pub fn into_resolution(self) -> Result<(i32, i32, i8), Self>

Returns the inner fields if this is a IppValue::Resolution, otherwise returns back the enum in the Err case of the result

pub fn as_no_value_mut(&mut self) -> Option<()>

Optionally returns references to the inner fields if this is a IppValue::NoValue, otherwise None

pub fn as_no_value(&self) -> Option<()>

Optionally returns references to the inner fields if this is a IppValue::NoValue, otherwise None

pub fn as_other_mut(&mut self) -> Option<(&mut u8, &mut Bytes)>

Optionally returns mutable references to the inner fields if this is a IppValue::Other, otherwise None

pub fn as_other(&self) -> Option<(&u8, &Bytes)>

Optionally returns references to the inner fields if this is a IppValue::Other, otherwise None

pub fn into_other(self) -> Result<(u8, Bytes), Self>

Returns the inner fields if this is a IppValue::Other, otherwise returns back the enum in the Err case of the result

impl IppValue

pub fn to_tag(&self) -> u8

Convert to binary tag

pub fn parse(value_tag: u8, data: Bytes) -> Result<IppValue>

Parse value from byte array which does not include the value length field

pub fn to_bytes(&self) -> Bytes

Write value to byte array, including leading value length field, excluding value tag